Fixed #122
