+++ src/main/scala-2.11/com/bwsw/tstreams/agents/consumer/subscriber/TransactionsBuffer.scala
@@ -2,8 +2,8 @@ package com.bwsw.tstreams.agents.consumer.subscriber
 
 import java.util.UUID
 
-import com.bwsw.tstreams.coordination.pubsub.messages.ProducerTransactionStatus
-import com.bwsw.tstreams.coordination.pubsub.messages.ProducerTransactionStatus._
+import com.bwsw.tstreams.coordination.messages.state.TransactionStatus._
+import com.bwsw.tstreams.coordination.messages.state.TransactionStatus
 
 /**
   * Buffer for maintaining consumed transactions in memory
@@ -23,19 +23,19 @@ class TransactionsBuffer {
 
     //TODO wrap all checks in validate method and log it
     //ignore update events until txn doesn't exist in buffer
-    if (!map.exist(txnUuid) && status == ProducerTransactionStatus.update) {
+    if (!map.exist(txnUuid) && status == TransactionStatus.update) {
       return
     }
 
     if (map.exist(txnUuid)) {
       map.get(txnUuid)._1 match {
-        case ProducerTransactionStatus.preCheckpoint =>
-          if (status != ProducerTransactionStatus.postCheckpoint &&
-            status != ProducerTransactionStatus.cancel) {
+        case TransactionStatus.preCheckpoint =>
+          if (status != TransactionStatus.postCheckpoint &&
+            status != TransactionStatus.cancel) {
             return
           }
 
-        case ProducerTransactionStatus.`postCheckpoint` =>
+        case TransactionStatus.`postCheckpoint` =>
           return
 
         case _ =>
@@ -43,19 +43,19 @@ class TransactionsBuffer {
     }
 
     status match {
-      case ProducerTransactionStatus.`update` |
-           ProducerTransactionStatus.opened =>
+      case TransactionStatus.`update` |
+           TransactionStatus.opened =>
         map.put(txnUuid, (status, ttl))
 
       //ignore ttl, preCheckpoint will be resolved by another thread
-      case ProducerTransactionStatus.preCheckpoint =>
+      case TransactionStatus.preCheckpoint =>
         map.put(txnUuid, (status, -1))
 
       //just ignore ttl because transaction is closed
-      case ProducerTransactionStatus.`postCheckpoint` =>
+      case TransactionStatus.`postCheckpoint` =>
         map.put(txnUuid, (status, -1))
 
-      case ProducerTransactionStatus.`cancel` =>
+      case TransactionStatus.`cancel` =>
         map.remove(txnUuid)
     }
   }
