+++ src/test/scala-2.11/agents/both/batch_insert/aerospike/ABasicProducerAndConsumerSimpleTests.scala
@@ -7,7 +7,7 @@ import com.bwsw.tstreams.agents.consumer.{Consumer, ConsumerOptions, ConsumerTra
 import com.bwsw.tstreams.agents.producer.DataInsertType.BatchInsert
 import com.bwsw.tstreams.agents.producer.{Producer, Options, CoordinationOptions, NewTransactionProducerPolicy}
 import com.bwsw.tstreams.common.CassandraHelper
-import com.bwsw.tstreams.coordination.transactions.transport.impl.TcpTransport
+import com.bwsw.tstreams.coordination.producer.transport.impl.TcpTransport
 import com.bwsw.tstreams.env.TSF_Dictionary
 import com.bwsw.tstreams.streams.TStream
 import org.scalatest.{BeforeAndAfterAll, FlatSpec, Matchers}
