+++ src/test/scala-2.11/common/ProtocolMessageSerializerTest.scala
@@ -1,10 +1,11 @@
 package common
 
 import java.util.UUID
+
 import com.bwsw.tstreams.common.ProtocolMessageSerializer
-import com.bwsw.tstreams.coordination.pubsub.messages.{ProducerTopicMessage, ProducerTransactionStatus}
-import com.bwsw.tstreams.coordination.transactions.messages._
-import com.bwsw.tstreams.coordination.transactions.peertopeer.AgentSettings
+import com.bwsw.tstreams.coordination.messages.master._
+import com.bwsw.tstreams.coordination.messages.state.{Message, TransactionStatus}
+import com.bwsw.tstreams.coordination.producer.p2p.AgentSettings
 import org.scalatest.{BeforeAndAfterAll, FlatSpec, Matchers}
 import testutils.TestUtils
 
@@ -49,13 +50,13 @@ class ProtocolMessageSerializerTest extends FlatSpec with Matchers with BeforeAn
     clazz shouldEqual req
   }
   "TStreams serializer" should "serialize and deserialize PublishRequest" in {
-    val clazz = PublishRequest("snd", "rcv", ProducerTopicMessage(UUID.randomUUID(), 228, ProducerTransactionStatus.cancel, 1488))
+    val clazz = PublishRequest("snd", "rcv", Message(UUID.randomUUID(), 228, TransactionStatus.cancel, 1488))
     val string = serializer.serialize(clazz)
     val req = serializer.deserialize[PublishRequest](string)
     clazz shouldEqual req
   }
   "TStreams serializer" should "serialize and deserialize PublishResponse" in {
-    val clazz = PublishResponse("snd", "rcv", ProducerTopicMessage(UUID.randomUUID(), 228, ProducerTransactionStatus.cancel, 1488))
+    val clazz = PublishResponse("snd", "rcv", Message(UUID.randomUUID(), 228, TransactionStatus.cancel, 1488))
     val string = serializer.serialize(clazz)
     val req = serializer.deserialize[PublishResponse](string)
     clazz shouldEqual req
@@ -85,22 +86,22 @@ class ProtocolMessageSerializerTest extends FlatSpec with Matchers with BeforeAn
     clazz shouldEqual req
   }
   "TStreams serializer" should "serialize and deserialize PTM" in {
-    val clazz = ProducerTopicMessage(UUID.randomUUID(), 123, ProducerTransactionStatus.postCheckpoint, 5)
+    val clazz = Message(UUID.randomUUID(), 123, TransactionStatus.postCheckpoint, 5)
     val string = serializer.serialize(clazz)
-    val req = serializer.deserialize[ProducerTopicMessage](string)
+    val req = serializer.deserialize[Message](string)
     clazz shouldEqual req
   }
   "TStreams serializer" should "serialize and deserialize PTS" in {
-    val pre = ProducerTransactionStatus.preCheckpoint
-    val fin = ProducerTransactionStatus.postCheckpoint
-    val canceled = ProducerTransactionStatus.cancel
-    val updated = ProducerTransactionStatus.update
-    val opened = ProducerTransactionStatus.opened
-    assert(serializer.deserialize[ProducerTransactionStatus.ProducerTransactionStatus](serializer.serialize(pre)) == pre)
-    assert(serializer.deserialize[ProducerTransactionStatus.ProducerTransactionStatus](serializer.serialize(fin)) == fin)
-    assert(serializer.deserialize[ProducerTransactionStatus.ProducerTransactionStatus](serializer.serialize(canceled)) == canceled)
-    assert(serializer.deserialize[ProducerTransactionStatus.ProducerTransactionStatus](serializer.serialize(updated)) == updated)
-    assert(serializer.deserialize[ProducerTransactionStatus.ProducerTransactionStatus](serializer.serialize(opened)) == opened)
+    val pre = TransactionStatus.preCheckpoint
+    val fin = TransactionStatus.postCheckpoint
+    val canceled = TransactionStatus.cancel
+    val updated = TransactionStatus.update
+    val opened = TransactionStatus.opened
+    assert(serializer.deserialize[TransactionStatus.ProducerTransactionStatus](serializer.serialize(pre)) == pre)
+    assert(serializer.deserialize[TransactionStatus.ProducerTransactionStatus](serializer.serialize(fin)) == fin)
+    assert(serializer.deserialize[TransactionStatus.ProducerTransactionStatus](serializer.serialize(canceled)) == canceled)
+    assert(serializer.deserialize[TransactionStatus.ProducerTransactionStatus](serializer.serialize(updated)) == updated)
+    assert(serializer.deserialize[TransactionStatus.ProducerTransactionStatus](serializer.serialize(opened)) == opened)
   }
   "TStreams serializer" should "serialize and deserialize AgentSettings" in {
     val clazz = new AgentSettings("agent", 21212, 12121212)
